Allow interpolation in clean
