+++ yapim/__init__.py
@@ -1,8 +1,9 @@
 """Yet Another PIpeline Manager"""
 from yapim.tasks.aggregate_task import AggregateTask
-from yapim.tasks.task import Task, clean
+from yapim.tasks.task import Task
 from yapim.tasks.task import TaskExecutionError
 from yapim.tasks.task import TaskSetupError
+from yapim.tasks.utils.clean import clean
 from yapim.tasks.utils.dependency_input import DependencyInput
 from yapim.tasks.utils.result import Result
 from yapim.tasks.utils.version_info import VersionInfo
